--- docs/reference/API/trestle/core/all_validator.md
@@ -1,7 +1,0 @@
----
-title: trestle.core.all_validator
-description: Documentation for trestle.core.all_validator module
----
-
-::: trestle.core.all_validator
-handler: python
