--- docs/reference/API/trestle/core/catalog_validator.md
@@ -1,7 +1,0 @@
----
-title: trestle.core.catalog_validator
-description: Documentation for trestle.core.catalog_validator module
----
-
-::: trestle.core.catalog_validator
-handler: python
