--- docs/reference/API/trestle/core/duplicates_validator.md
@@ -1,7 +1,0 @@
----
-title: trestle.core.duplicates_validator
-description: Documentation for trestle.core.duplicates_validator module
----
-
-::: trestle.core.duplicates_validator
-handler: python
