--- docs/reference/API/trestle/core/links_validator.md
@@ -1,7 +1,0 @@
----
-title: trestle.core.links_validator
-description: Documentation for trestle.core.links_validator module
----
-
-::: trestle.core.links_validator
-handler: python
