--- docs/reference/API/trestle/core/refs_validator.md
@@ -1,7 +1,0 @@
----
-title: trestle.core.refs_validator
-description: Documentation for trestle.core.refs_validator module
----
-
-::: trestle.core.refs_validator
-handler: python
