--- docs/reference/API/trestle/core/rule_parameters_validator.md
@@ -1,7 +1,0 @@
----
-title: trestle.core.rule_parameters_validator
-description: Documentation for trestle.core.rule_parameters_validator module
----
-
-::: trestle.core.rule_parameters_validator
-handler: python
