--- docs/reference/API/trestle/core/validator.md
@@ -1,7 +1,0 @@
----
-title: trestle.core.validator
-description: Documentation for trestle.core.validator module
----
-
-::: trestle.core.validator
-handler: python
