--- docs/reference/API/trestle/core/validator_factory.md
@@ -1,7 +1,0 @@
----
-title: trestle.core.validator_factory
-description: Documentation for trestle.core.validator_factory module
----
-
-::: trestle.core.validator_factory
-handler: python
