+++ docs/reference/API/trestle/sdk/validation/catalog_validator.md
@@ -1,0 +1,7 @@
+---
+title: trestle.sdk.validation.catalog_validator
+description: Documentation for trestle.sdk.validation.catalog_validator module
+---
+
+::: trestle.sdk.validation.catalog_validator
+handler: python
