+++ docs/reference/API/trestle/sdk/validation/duplicates_validator.md
@@ -1,0 +1,7 @@
+---
+title: trestle.sdk.validation.duplicates_validator
+description: Documentation for trestle.sdk.validation.duplicates_validator module
+---
+
+::: trestle.sdk.validation.duplicates_validator
+handler: python
