+++ docs/reference/API/trestle/sdk/validation/refs_validator.md
@@ -1,0 +1,7 @@
+---
+title: trestle.sdk.validation.refs_validator
+description: Documentation for trestle.sdk.validation.refs_validator module
+---
+
+::: trestle.sdk.validation.refs_validator
+handler: python
