+++ trestle/sdk/validation/duplicates_validator.py
@@ -18,7 +18,7 @@ import pathlib
 from typing import Optional
 
 from trestle.common.model_utils import ModelUtils
-from trestle.core.validator import Validator
+from trestle.sdk.validation.validator import Validator
 from trestle.sdk.models.base_model import OscalBaseModel
 from trestle.sdk.oscal.profile import Profile
