+++ trestle/sdk/validation/refs_validator.py
@@ -19,7 +19,7 @@ from typing import Optional
 
 from trestle.common.common_types import TopLevelOscalModel
 from trestle.common.list_utils import as_list
-from trestle.core.validator import Validator
+from trestle.sdk.validation.validator import Validator
 
 
 class RefsValidator(Validator):
