+++ trestle/sdk/validation/rule_parameters_validator.py
@@ -23,7 +23,7 @@ from trestle.common.common_types import TypeWithSetParams
 from trestle.common.list_utils import as_list, deep_set
 from trestle.core.catalog.catalog_interface import CatalogInterface
 from trestle.core.profile_resolver import ProfileResolver
-from trestle.core.validator import Validator
+from trestle.sdk.validation.validator import Validator
 from trestle.sdk.oscal.ssp import ImplementedRequirement, SystemSecurityPlan
 
 logger = logging.getLogger(__name__)
